use template for XXX EventQUeue

+++ sora/celshading_main.cpp
@@ -83,8 +83,8 @@ namespace celshading {
   }
 
   void setup_graphics(sora::Device *dev, int w, int h) {
-    win_width = w;
-    win_height = h;
+    win_width = (float)w;
+    win_height = (float)h;
 
     //edge 그리기
     {
@@ -142,7 +142,7 @@ namespace celshading {
     //obj rendering
     glClearColor(0.3f, 0.8f, 0.8f, 1.0f);
     glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-    glViewport(0, 0, win_width, win_height);
+    glViewport(0, 0, (int)win_width, (int)win_height);
     
 
     //edge 그리기
