+++ sora/depth_map_main.cpp
@@ -68,8 +68,8 @@ namespace depthmap {
   GLuint color_rb = 0;
 
   void setup_graphics(sora::Device *dev, int w, int h) {
-    win_width = w;
-    win_height = h;
+    win_width = (float)w;
+    win_height = (float)h;
 
     //create shader
     {
@@ -145,7 +145,7 @@ namespace depthmap {
       
       glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
       glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-      glViewport(0, 0, win_width, win_height);
+      glViewport(0, 0, (int)win_width, (int)win_height);
 
       Renderer &render3d = dev->render3d();
       render3d.SetShader(simple_shader);
@@ -183,7 +183,7 @@ namespace depthmap {
     {
       glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
       glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-      glViewport(0, 0, win_width, win_height);
+      glViewport(0, 0, (int)win_width, (int)win_height);
 
       //깊이맵을 적절히 렌더링
       float vertex[] = {
